Changed type of data to float.

+++ LaparoCommunicator/LaparoCommunicator/Communicators/LaparoCommunicatorMock.cs
@@ -27,10 +27,10 @@ namespace LaparoCommunicator
                 position = 0;
 
             var data = eulerData[position];
-            var leftCoords = data.Split(CoordSeparator, StringSplitOptions.RemoveEmptyEntries).Select(double.Parse).ToArray();
+            var leftCoords = data.Split(CoordSeparator, StringSplitOptions.RemoveEmptyEntries).Select(float.Parse).ToArray();
 
             data = eulerData[position + 1];
-            var rightCoords = data.Split(CoordSeparator, StringSplitOptions.RemoveEmptyEntries).Select(double.Parse).ToArray();
+            var rightCoords = data.Split(CoordSeparator, StringSplitOptions.RemoveEmptyEntries).Select(float.Parse).ToArray();
             position = position + 2;
 
             return new EulerData() { LeftAngles = leftCoords, RightAngles = rightCoords };
@@ -42,10 +42,10 @@ namespace LaparoCommunicator
                 position = 0;
 
             var data = quaternionData[position];
-            var leftCoords = data.Split(CoordSeparator, StringSplitOptions.RemoveEmptyEntries).Select(double.Parse).ToArray();
+            var leftCoords = data.Split(CoordSeparator, StringSplitOptions.RemoveEmptyEntries).Select(float.Parse).ToArray();
 
             data = quaternionData[position + 1];
-            var rightCoords = data.Split(CoordSeparator, StringSplitOptions.RemoveEmptyEntries).Select(double.Parse).ToArray();
+            var rightCoords = data.Split(CoordSeparator, StringSplitOptions.RemoveEmptyEntries).Select(float.Parse).ToArray();
             position = position + 2;
 
             return new QuaternionData() { LeftQuaternion = leftCoords, RightQuaternion = rightCoords };
